+++ Card/src/androidTest/java/com/braintreepayments/api/card/CardClientTest.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.api;
+package com.braintreepayments.api.card;
 
 import static com.braintreepayments.api.CardNumber.VISA;
 import static com.braintreepayments.api.Fixtures.TOKENIZATION_KEY;
@@ -10,6 +10,17 @@ import static junit.framework.Assert.assertTrue;
 import android.content.Context;
 
 import androidx.test.core.app.ApplicationProvider;
+
+import com.braintreepayments.api.Authorization;
+import com.braintreepayments.api.AuthorizationException;
+import com.braintreepayments.api.BraintreeClient;
+import com.braintreepayments.api.Configuration;
+import com.braintreepayments.api.ErrorWithResponse;
+import com.braintreepayments.api.Fixtures;
+import com.braintreepayments.api.SharedPreferencesHelper;
+import com.braintreepayments.api.TestClientTokenBuilder;
+
+import junit.framework.Assert;
 
 import org.json.JSONException;
 import org.junit.Ignore;
@@ -197,7 +208,7 @@ public class CardClientTest {
         sut.tokenize(card, (cardResult) -> {
             assertTrue(cardResult instanceof CardResult.Failure);
             Exception error = ((CardResult.Failure) cardResult).getError();
-            assertEquals("CVV verification failed",
+           assertEquals("CVV verification failed",
                     ((ErrorWithResponse) error).errorFor("creditCard").getFieldErrors().get(0).getMessage());
             countDownLatch.countDown();
         });