+++ Card/src/main/java/com/braintreepayments/api/card/Card.java
@@ -1,4 +1,4 @@
-package com.braintreepayments.api;
+package com.braintreepayments.api.card;
 
 import android.os.Parcel;
 import android.os.Parcelable;
@@ -7,6 +7,7 @@ import android.text.TextUtils;
 import androidx.annotation.Nullable;
 import androidx.annotation.RestrictTo;
 
+import com.braintreepayments.api.BraintreeException;
 import com.braintreepayments.api.GraphQLConstants.Keys;
 
 import org.json.JSONException;
@@ -28,7 +29,8 @@ public class Card extends BaseCard implements Parcelable {
 
     private boolean shouldValidate;
 
-    JSONObject buildJSONForGraphQL() throws BraintreeException, JSONException {
+    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+    public JSONObject buildJSONForGraphQL() throws BraintreeException, JSONException {
         JSONObject base = new JSONObject();
         JSONObject input = new JSONObject();
         JSONObject variables = new JSONObject();