+++ Card/src/main/java/com/braintreepayments/api/card/CardClient.java
@@ -1,9 +1,15 @@
-package com.braintreepayments.api;
+package com.braintreepayments.api.card;
 
 import android.content.Context;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
+
+import com.braintreepayments.api.ApiClient;
+import com.braintreepayments.api.BraintreeClient;
+import com.braintreepayments.api.BraintreeException;
+import com.braintreepayments.api.ErrorWithResponse;
+import com.braintreepayments.api.GraphQLConstants;
 
 import org.json.JSONException;
 import org.json.JSONObject;