+++ Card/src/main/java/com/braintreepayments/api/card/CardTokenizeCallback.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.card
 
 /**
  * Callback for receiving result of [CardClient.tokenize].